backend/requirements.txt
django==5.0.4
environs==11.0.0
psycopg2-binary==2.9.9
django-allauth==64.0.0
dj-rest-auth==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-countries[pyuca]==7.6.1
django-environ
cryptography
django-cryptography